requirements.txt
python-telegram-bot
python-dotenv
qrcode
json
os
datetime